Move button release handling to seperate event helper
